RUNLOOP: Allow to adjust the timer clock prescaler during initialization. Add support for error callbacks.
The runloop subsystem now allows to register a task error callback and a synchronization error callback.
The first is executed if a task returns with an error code other than RUNLOOP_OK or RUNLOOP_OK_TASK_ABORT.
The latter is executed if the execution time of a task delays another task by more than its execution period.
In this situation, the runloop drops task executions.