RUNLOOP: Add uptime functionality.
The uptime functionality can be enabled by the new macro RUNLOOP_WITH_UPTIME.
When enabled, there are two new functions: RUNLOOP_GetUptimeClockCycles()
and RUNLOOP_GetUptimeHumanReadable().
The macro RUNLOOP_UPTIME_UPDATE_INTERVAL_MS allows to set a minimum uptime update interval.
This is meant to be used in case that the uptime is queried from interrupt service routines
while the runloop is idle.