RUNLOOP: Integrate new CMDL interface for better interoperability.
CMDL is now automatically integrated into the RUNLOOP if the macro
RUNLOOP_WITH_CMDL is set. The RUNLOOP now allows to execute CMDL
commands while tasks are running in the background. In addition, the
RUNLOOP can be paused by the "pause" command. If CMDL is not integrated,
the runloop can be paused by pressing the 'q' key.