RUNLOOP test: Adjust for new CMDL and RUNLOOP interface + improvements.
The application can be compiled with RUNLOOP_WITH_CMDL both set and
unset. If set, tasks can be added to the runloop interactively while the
runloop is running or paused. Otherwise, some keys can be used to add
tasks with default values.
The reset source is printed after application startup.
The watchdog is handled correctly after a hardware reset now.